拉取店铺订单同时同步到OMS订单库

+++ api/src/main/java/cn/qihangerp/api/service/ShopOrderService.java
@@ -13,6 +13,6 @@ import com.baomidou.mybatisplus.extension.service.IService;
 */
 public interface ShopOrderService extends IService<ShopOrder> {
     PageResult<ShopOrder> queryPageList(ShopOrder bo, PageQuery pageQuery);
-    ResultVo<Integer> saveOrder(Long shopId, ShopOrder order);
+    ResultVo<Long> saveOrder(Long shopId, ShopOrder order);
     ResultVo<Integer> orderConfirm(String[] ids);
 }
